Use multi-staged Docker builds.
We've been using a nested Docker builds as a way to produce small
runtime images that don't have to include the whole build environment.
This is kind of messy and requires a `docker build` and `docker run`
cycle just to produce the final image.

However Docker provides that feature natively through multi-staged
builds: https://docs.docker.com/build/building/multi-stage/
This simplifies the pipeline slightly, and the image may now be built as
a single step.