mrc-4765: Integrate the emulator into the frontend.
This adds a new dropdown in the title bar, which may be use to choose
between the pre-calculated dataset or one of the available emulator
models.

The emulator model is loaded into the browser as an ONNX file, and
executed using the onnxruntime library. The frontend makes little
assumption about the shape of the inputs and how they need to be
processed into a tensor. Instead it uses a JSON description of the
features, that is served by mintr and can be modified in sync with the
form options.

If mintr was not started with any emulators, the drop-down is hidden and
the frontend operates as before, using the precalculated data only.